Duplicate Unsplash enum This removes a cross reference across the two extension contexts. (Shell background and preferences window)
